construccion de DTO para la manipulacion de las entidades en el swagger

+++ Proyecto2025.Server/Proyecto2025.Server/Controllers/MessageController.cs
@@ -57,6 +57,10 @@ namespace Proyecto2025.Server.Controllers
                 return NotFound($"No Existe el Mensaje con el Id: {Id}.");
             }
             var messaje = await context.Messages.FirstOrDefaultAsync(x => x.Id == Id);
+            if (messaje is null)
+            {
+                return NotFound($"No se encontro el Mensaje con el Id: {Id}.");
+            }
             context.Messages.Remove(messaje);
             await context.SaveChangesAsync();
             return Ok($"Mensaje con el Id {Id} eliminado correctamente");
